+++ CHANGELOG.md
@@ -8,9 +8,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Features
 - Update Colors using the `System.Windows.Media.Colors` as reference.
 - Update `MaterialUtils` and `GraphicsStyleUtils` with default 9 colors.
+- Add `ColorExtension` with `ColorEquals`, `Lerp`, `ToColor` and `ToColorWithTransparency`.
+- Add `ColorExtension` with `ToHex` for `Color` and `ColorWithTransparency`
 ### Tests
 - Test all color names in `System.Windows.Media.Colors` with `Colors`.
 - Test all 9 colors in `MaterialUtils` and `GraphicsStyleUtils`.
+- Test for `ColorExternsion`
 
 ## [0.3.1] / 2024-01-09 - 2024-04-13
 ### Features